Add more type-safe list
At the expense of a lot of verbosity. Not sure if there's a better way, but it's challenging because lists can have all sorts of different data types, AND are nested Cons cells.

Also note that because lists can theoretically be the empty list, I have to handle that case in `getHead` and `getTail`.